Make the service provider draw the frustum
And rotate the frustum for the VR Service provider
