+++ Assets/LeapMotion/Editor/LeapVRServiceProviderEditor.cs.meta
@@ -1,8 +1,12 @@
 fileFormatVersion: 2
-guid: e617762d9583845b884c0dce6e2b54a6
+guid: a265067a2614ee14bac2bbac11148a5d
+timeCreated: 1494628409
+licenseType: Pro
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
   userData: 
+  assetBundleName: 
+  assetBundleVariant:
